Deleted pyc files and renamed code file
